restructuring and clean up for autopooling and termination
