+++ src/rust/WHATIVELEARNED
@@ -4,3 +4,5 @@
 - async vs blocking operations
 - reverse proxying
 - difference between redirect and proxy
+- i sort of understand what a Cell is
+- I definitely understand Arc and Mutex now
